+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - Add meeting demo parameter for broadcasting user
 - Add simulcast guide link in README and Quality Bandwidth Connectivity doc
+- Add a MediaDevices proxy to support the devicechange event in some Android Chrome browsers
 
 ### Changed